+++ src/Sentry/Integrations/AppDomainUnhandledExceptionIntegration.cs
@@ -6,7 +6,7 @@ using Sentry.Protocol;
 
 namespace Sentry.Integrations
 {
-    internal class AppDomainUnhandledExceptionIntegration : IInternalSdkIntegration
+    internal class AppDomainUnhandledExceptionIntegration : ISdkIntegration
     {
         private readonly IAppDomain _appDomain;
         private IHub? _hub;
@@ -20,12 +20,6 @@ namespace Sentry.Integrations
             _appDomain.UnhandledException += Handle;
         }
 
-        public void Unregister(IHub hub)
-        {
-            _appDomain.UnhandledException -= Handle;
-            _hub = null;
-        }
-
         // Internal for testability
 #if !NET6_0_OR_GREATER
         [HandleProcessCorruptedStateExceptions]
@@ -37,7 +31,7 @@ namespace Sentry.Integrations
             {
                 ex.Data[Mechanism.HandledKey] = false;
                 ex.Data[Mechanism.MechanismKey] = "AppDomain.UnhandledException";
-                _ = (_hub?.CaptureException(ex));
+                _ = _hub?.CaptureException(ex);
             }
 
             if (e.IsTerminating)